tests, and some attempt to clean up the mess
There's now a Makefile to create expander.js. You need to have `pegjs`
available at the command line. Install it with `npm install -g pegjs`
(this requires npm, which requires node, etc.)

test.html runs a series of unit tests to verify that the base grammar
works the same as the Numbas JME parser. At the moment, it fails in a
lot of cases.